+++ CHANGES.rst
@@ -4,6 +4,12 @@ Changelog
 1.0rc2 (unreleased)
 -------------------
 
+- Allow limiting collection results (closes `#6`_).
+  [hvelarde]
+
+- Remove dependency on five.grok.
+  [hvelarde]
+
 - Drop support for Python 2.6 and Plone 4.1;
   Remove dependency on unittest2.
   [hvelarde]
@@ -19,3 +25,4 @@ Changelog
 
 - Initial release.
 
+.. _`#6`: https://github.com/collective/collective.newsticker/issues/6